# packages/treeai-sdk/openapi/inference.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AI Inference Control API",
    "version": "0.11.0"
  },
  "paths": {
...
    "/v1/events/stream": {
      "get": {
        "operationId": "inference.get.events.stream",
        "summary": "Stream job changes",
        "security": [
          {
            "apiKey": [
              "jobs:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_<id>_<secret>"
      }
    }
  }
}